+++ src/network/utils/inet-socket-address.cc
@@ -120,7 +120,7 @@ uint8_t
 InetSocketAddress::GetType()
 {
     NS_LOG_FUNCTION_NOARGS();
-    static uint8_t type = Address::Register();
+    static uint8_t type = Address::Register("IpAddress", 6);
     return type;
 }