+++ src/network/utils/ipv4-address.cc
@@ -361,7 +361,7 @@ uint8_t
 Ipv4Address::GetType()
 {
     NS_LOG_FUNCTION_NOARGS();
-    static uint8_t type = Address::Register();
+    static uint8_t type = Address::Register("IpAddress", 4);
     return type;
 }