+++ src/network/utils/ipv6-address.cc
@@ -681,7 +681,7 @@ uint8_t
 Ipv6Address::GetType()
 {
     NS_LOG_FUNCTION_NOARGS();
-    static uint8_t type = Address::Register();
+    static uint8_t type = Address::Register("IpAddress", 16);
     return type;
 }